# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "proviant is a simple and intuitive application to track your bought products and their expiration date to prevent waste of food.",
        "title": "proviant",
        "contact": {
            "name": "Isotop7",
            "url": "https://codeberg.org/isotop7/proviant",
            "email": "hendrik@hr94.de"
        },
        "license": {
            "name": "MIT",
            "url": "https://codeberg.org/isotop7/proviant/-/blob/main/LICENSE"
        },
        "version": "0.2.0"
    },
    "host": "localhost:5050",
    "basePath": "/",
    "paths": {
...
        "/api/v1/calendar/export.ics": {
            "get": {
                "description": "Returns an iCalendar (RFC 5545) feed with VEVENTs for products expiring in the next 30 days. Use the token query parameter for authentication.",
                "produces": [
                    "text/calendar"
                ],
                "tags": [
                    "calendar"
                ],
                "summary": "Export iCalendar feed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Calendar token for authentication",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "iCalendar feed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JSON Web Token": {
            "type": "basic"
        }
    },
...
}